+++ index.js
@@ -3,7 +3,8 @@ var readDocumentData = require('./lib/read-document-data')
 var removeDocumentData = require('./gulp/remove-document-data')
 var loadDocumentData = require('./gulp/load-document-data')
 
-var articleList = require('./helper/articleList')
+var articleList = require('./helper/article/list')
+var articleSticky = require('./helper/article/sticky')
 var markdown = require('./helper/markdown')
 var tagNavigation = require('./helper/tagNavigation')
 var formatTags = require('./helper/formatTags')
@@ -22,6 +23,7 @@ module.exports = {
     tagNavigation: tagNavigation,
     formatTags: formatTags,
     articleList: articleList,
+    articleSticky: articleSticky,
     markdown: markdown
   }
 }