+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.3.0
+- FEATURE: Add Sticky feature for article list
+- BREAKING: You have to add the parameter for respectSticky(true|false) on the articleList helper call
+
 ## 1.2.1
 - Removed wrong badges from readme
 - Migrated from `extend` to `Object.assign()`